Python lib for Deep Reinforcement Learning
Currently use the openai gym as the environment to test the learning algorithm. In this case, the openai gym Ros wrapper is built to test the communication between the read world and the learning agent.The framework of the openai gym Ros wrapper would refer as the prototype for building a large scale simulation base on Gazebo.